rgw: false assumption on vault bucket key deletion
On bucket key deletion when the request to change the property of
the key for deletion_allowed to true, it is expected that the
response body be empty. But this assumption is false and there
would be a dump of the new config in the response. this condition
would prevent the key deletion to be done.

Fixes: https://tracker.ceph.com/issues/65626
Signed-off-by: Seena Fallah <seenafallah@gmail.com>
(cherry picked from commit 094bfc535eb0b54b1d4fcff3931affdb2c6e7785)