mgr: do not require NOTIFY_TYPES in python modules
Many python mgr modules lack a NOTIFY_TYPES member apparently without
any downsides beyond an annoying message in the logs.  This includes
commonly used mgr modules like 'volumes' and 'cephadm'.  Stop emitting
an error for missing NOTIFY_TYPES by not assuming all mgr modules in
python should provide it. The return code is also changed to indicate
that this is not an error but the return from this function is never
checked and so it should have no impact. However, if someone started
using the return value in the future this would match the log change.

Fixes: https://tracker.ceph.com/issues/55835
Signed-off-by: John Mulligan <jmulligan@redhat.com>