mgr/cephadm: add retry when running ganesha-rados-grace tool
This consistently seems to hit failures while running it

ganesha-rados-grace tool failed: rados_pool_create: -1
Can't connect to cluster: -1
terminate called after throwing an instance of 'std::bad_variant_access'
  what():  std::get: wrong index for variant

but on follow up attempts seems to work okay and eventually
the nfs daemon we're deploying seems to come up alright. In
the meantime a CEPHADM_DAEMON_PLACE_FAIL is raised for the
nfs daemon, that is then cleared up on its own when cephadm
tries to deploy the daemon again on its next iteration of
the serve loop. This adds a retry to the run_grace_tool
function in the hope we can just do the retry immediately
rather than raising a health warning for failing to deploy
a daemon and then retrying later.

Fixes: https://tracker.ceph.com/issues/65702

Signed-off-by: Adam King <adking@redhat.com>