qa/cephadm: ignore CEPHADM_STRAY_DAEMON in upgrade tests
There is some other ongoing work to make it so daemons
created/destroyed in the last minute are not marked stray
as there can be a delay between when cephadm considers the
daemon as being removed and when the core mgr reports the
same. Unfortunately, that change only fixes it for releases
in which the patch will be present, so the upgrade test
is still susceptible to having these warnings pop up
before the upgrade occurs.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit c4a875f6810923e0efb5ed752b5df3ec1b2f4633)