qa/config/crimson_qa_overrides: adjust mgr_stats_period
Some test jobs fail with `local_shared_foreign_ptr: Assertion `ptr && *ptr' failed`
It seems that we attempt to use a connection which is not yet ready to use
after setting up the daemons on boot.
Adjust the mgr_stats_period to allow more time for the daemons to set up.
See: https://tracker.ceph.com/issues/62162#note-10

Note: This is not a fix but more of a temporary solution to avoid noise
      in the testing suite (Tracker stays open).

Signed-off-by: Matan Breizman <mbreizma@redhat.com>