osd/PrimaryLogPG: fix LOST_REVERT on primary
Commit 3a0467749b6f27a21483de304fde46b03c3836ca introduced pulling
old version to revert to in LOST_REVERT:
```
osd: pull old version to revert to
If we are the primary, and are doing a LOST_REVERT, pull the old version
of the object and update the version when we get it.
```
A dedicated `revert` boolean was used under the `if (is_primary())` scope.

Commit 1bf037bf76d0c3f85421657a5aaf8b7478937e02 refactored the push and pull
code. The refactor dropped the `revert` boolean and instead asserted that
we are primary. The assertion was correct since this bit was moved to `submit_push_complete`
which was called in `handle_pull_response` that was called from `sub_op_push` only on *primary*.

Commit 7da308bba0d60b1435cda9698a3320d97050699b moved this logic into the common `on_local_recover`.
Now `on_local_recover` was called from `handle_pull_response` not only on *primary*.
Therfore, `on_local_recover` included a `is_primary` check in the end since it was called also from
non primaries.

However, the LOST_REVERT logic should be used only on primary so we need to bring back the `revert`
boolean same as 3a0467749b6f27a21483de304fde46b03c3836ca did.

Fixes:  https://tracker.ceph.com/issues/50608

Signed-off-by: Matan Breizman <mbreizma@redhat.com>