mgr/smb: add login_control and restrict_access attributes to share
The login_control attribute contains a list of LoginAccessEntry objects.
These will be mapped to samba share access parameters.
The restrict_access attribute is a boolean that makes access to the
share contingent on being listed in the login_control list, as
either a listed user or member of a listed group, for any access
level other than 'none' (as that always denies access).

Signed-off-by: John Mulligan <jmulligan@redhat.com>