mgr/nfs: preserve user_id validations and fix its generation
Signed-off-by: Avan Thakkar <athakkar@redhat.com>

The fix keeps the same user_id and cephx_key generation the same, no need
to take into account fs name or cmount path because that will not handle the case
well when we create export with same cmount_path & fs_name  but different `path`
for e.g. subvolume path & if we delete that export it'll also delete the client
which the other export would be consuming. So taking into account the export_id &
path should be enough for generation of user_id & cephx_key attached to it.