qa/ceph_test_case: check for error in stderr in run_ceph_cmd
Idempotent commands like "ceph mds fail" return zero (for the sake of
idempotency) but also print error message on stderr when a non-existent
MDS's name is passed to it.

Now in QA code, if a non-existent MDS's name is passsed by mistake to
"ceph mds fail", it goes unnoticed because the command passes instead of
failing. This in turns ensure tests finishes successfully instead of
crashing and halting.

To avoid such issues, check for errors in stderr and if it/they exist,
raise CommandFailedError if check_status is set to True.

Signed-off-by: Rishabh Dave <ridave@redhat.com>