RGW: Remove get_obj_state()/set_obj_state from SAL
RGWObjState is the state for the StoreObject class.  It has historically
been accessible via get_obj_state()/set_obj_state(), but the double
pointer nature of this access has caused multiple bugs, and the
RGWObjState itself is an implementation detail that doesn't need to be
exposed.

Instead, add a load_obj_state() that loads the state from the store, and
use proper getters/setters for the data.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
Signed-off-by: matt benjamin <mbenjamin@redhat.com>