mds: Convert referent_inode list to map
When multi-mdses are involved in hardlink creation and
if the mds which is driving the hardlink operation crashes
after sending link_prep msg to mds (real auth) of primary
inode, the rollback kicks when the mds comes back and fixes
the nlink and removes the referent inode. But before the
rollback kicks in, a lookup can come on primary inode on
the mds (real auth), which fails to load the referent inode
as it's not commited to back end yet.

The map helps to track these intransit operations. The referent
inode is usable only when [referent_inode] = 1 and not when
[referent_inode] = 0.

TODO:
 Revisit - Can't this be fixed by adding the referent inode to
the list at the end where it's marked as [referent_inode] = 1 ???
This sounds more simpler.

Signed-off-by: Kotresh HR <khiremat@redhat.com>