mgr/dashboard: apply replication policy for a bucket
On a normal multisite configured cluster, you can create a bucket with
this replication enabled which will stop the normal syncing and starts
doing the granular bucket syncing; meaning only the bucket with the
replication enabled will be syncing to the secondary site.

To enable replication, there should be a group policy created in the
primary site. If no group policy is there, the dashboard will create
one with bidirectional rule and add all the zones in the zonegroup for
syncing.

Fixes: https://tracker.ceph.com/issues/66239
Signed-off-by: Nizamudeen A <nia@redhat.com>