doc/cephfs: add "metrics" to left pane
Add the "Metrics" document to the Quincy branch. This commit brings the
Quincy branch up-to-date with the changes made in 56c7cc8c593.

Cf. https://github.com/ceph/ceph/pull/53003, which mixed docs updates
with code updates and therefore could not at that time cleanly be
backported into the docs (by me).

Signed-off-by: Zac Dover <zac.dover@proton.me>