cmake: fix the bug of compiling with rdma
When ceph needs to use the rdma function, the rdma compilation option
needs to be turned on. However, once the rdma compilation option is
turned on, ceph cannot be compiled. After analyzing the problem, it is
found that there is a problem with a compilation file of ceph, this
patch is for fix this bug.

Signed-off-by: Xiaobing Li <xiaobing.li@samsung.com>