doc/start: s/http/https/ in links
Replace "http" with "https" in doc/start/get-involved.rst.

This commit is, in a way, a repeat of
https://github.com/ceph/ceph/pull/57213/
(1c5383b91bd7dbfa9670c6485fcc5ff28b79f40d), which targeted the Reef
branch instead of the main branch. When this commit has been merged and
backported, I will close https://github.com/ceph/ceph/pull/57213/.

I am listing Casey Cain here as the co-author, but he is in fact the
true author of this change.

Co-authored-by: Casey Cain <ccain@linuxfoundation.org>
Signed-off-by: Zac Dover <zac.dover@proton.me>