mgr/cephadm: moving to a modular approach to manage nginx config
so far we have been using only one file with all the
configuration. This has the benefit of maintaining everything within
the same file but it can be really complex when more advanced
configuration is added for authentication for example. The new
approach splits the configuration into three files: main, external and
internal server configuration for better maintainability.

Signed-off-by: Redouane Kachach <rkachach@ibm.com>