mgr/dashboard: select default daemon based on the default zonegroup
if multisite is configured, the default daemon needs to be selected
based on the default zonegroup. Otherwise dashboard gives you incorrect
details when doing the period commit

The issue occurs when you do a period update --commit and you reload one
of the block page, the api assigns the zonegroup of the second gateway
because for a moment, the first gateway reflects the period changes...
This is not true because the default zonegroup is of the previous active
gateway but even though the back-end correctly says the active
zonegroup, the dashboard api says it wrongly.

Fixes: https://tracker.ceph.com/issues/66394
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit 013d0826c12e4bfc9dc2b8984b1605694ac7d2ea)