AsyncMessenger:  l_msgr_active_connections numerical anomaly
The counter (msgr_active_connections) can be an anomaly in case
if the counter is decrese before increase and initial value is 0. It can be
happen while the server daemon is blocked on accept_conn and client sends
a disconnect request.To avoid the situation increase the counter at first
step in add_accept during accepting a request so that the counter would not
be 0 during the decrease operation.

Fixes: https://tracker.ceph.com/issues/66231
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>
(cherry picked from commit b931251b2594eec601d39771b731c14119995181)