qa/suites/upgrade/telemetry-upgrade: upgrade from reef instead of pacific
With cephadm upgrades, we are only allowed to upgrade from as far back as N-2
releases. On the main branch, that means we can only upgrade from quincy and reef, and
we can no longer upgrade from pacific.

This test was trying to upgrade from pacific, which isn't allowed, which led to an
`UPGRADE_BAD_TARGET_VERSION` cluster error.

Fixes: https://tracker.ceph.com/issues/65189
Signed-off-by: Laura Flores <lflores@ibm.com>