cmake: add legacy-option-headers' sources with INTERFACE
to address the failure when generating the building system like:

```
CMake Error at src/common/options/CMakeLists.txt:108 (target_sources):
  target_sources may only set INTERFACE properties on INTERFACE targets
```

let's add `legacy-option-headers`'s source files with `INTERFACE`.
also, in order to workaround
https://gitlab.kitware.com/cmake/cmake/-/issues/18399, we have to set
the GENERATED property of the generated header file in very CMakeLists.txt,
so that it's visible to all cmake script where the dependency should be populated.
so a helper function of `add_ceph_options()` is added.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>