ceph_test_rados_api_misc: adjust LibRadosMiscConnectFailure.ConnectTimeout timeout
The test sets a 2 second connection timeout and asserts that the
attempt takes less than 4 seconds.

Under load, the asserted timeout can be exceeded. In particular,
this affects the Windows CI job, which uses a VM with a reduced
amount of resources.

    [ RUN      ] LibRadosMiscConnectFailure.ConnectTimeout
    /home/ubuntu/ceph/src/test/librados/misc.cc:87: Failure
    Expected: (dur) < (utime_t(4, 0)), actual: 5.987712 vs 4.000000

We'll adjust the timeouts, using a 5s configured timeout, asserting
that the connection attempt doesn't take more than 15s.

Fixes: https://tracker.ceph.com/issues/66534

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>