crush: fix the unreasonable weight_set poniter handling
This pr fix the bug which caused by unreasonable pointer handling. The weight_set pointer in crush_choose_arg struct may be a non-null pointer although the weight_set_positions is 0 when init, but ceph assert the weight_set_positions > 0 when the pointer is not null before use it.

Fixs: https://tracker.ceph.com/issues/66662
Signed-off-by: wanghai3 <wanghai3@360.cn>