mds: do not allow acquire wrlock for loner client requests
If the 'Fx' caps is issued the client will cache the dirty metadatas,
while if a write request, mkdir request for example, comes it will
try to update the parents' mtime without trigger to revoke the
corresponding caps. This will cause the mtime will be overwrote
later after the dirty caps is flushed from clients.

So for loner client's requests we need to revoke the 'Fx' caps anyway.

Fixes: https://tracker.ceph.com/issues/61584
Signed-off-by: Xiubo Li <xiubli@redhat.com>