mgr/cephadm: stop using "self.gws" in nvmeof service class
Using some "self.<attr>" in this class for anything that isn't constant
doesn't work. For one, on any restart of the cephadm mgr module,
including on mgr failover, upgrade, etc. will cause this attribute to
be lost. Furthermore, this makes the class unable to handle multiple
nvmeof services. If you had multiple services before this patch
and then removed one of them, assuming "self.gws" hasn't already
been dropped due to the cephadm module restarting, it would
cause ALL nvmeof daemons in the cluster to be run through the
"nvme-gw delete" command, not just the ones tied to the service
being removed. The most sane way to handle this removal imo is
to do so when the daemon is being removed. The extra attempt to
run all the nvmeof daemons through "nvme-gw delete" when purging
the service is unnecessary and complicates the process. We
should rely on this happening in "post_remove" which handles
it for individual daemons and if that doesn't work we should
fix it rather than trying to just do it again when removing
the service.

Resolves: rhbz#2294382

Signed-off-by: Adam King <adking@redhat.com>