script/ceph-backport: do not assume ref layout for checkout
The local repository may have a different layout from the defaults
for remote repository references. As for myself, I have:

    [remote "upstream"]
        url = git@github.com:ceph/ceph.git
        fetch = +refs/*:refs/remotes/upstream/*

So I have a local mirror of the full ceph.git repsository and upstream branches are
available via `refs/remotes/upstream/heads/<branch>`.

So, fetch the desired branch and use FETCH_HEAD instead. This is not 100% the same
as the merge branch for the wip- branch becomes:

    [branch "wip-66572-squid"]
        remote = origin
        merge = refs/heads/wip-66572-squid

So the remote tracking branch is now the origin. Rebasing to a newer <release> branch
must be done manually:

    git rebase upstream/heads/<release>

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>