mgr/rgw: fix setting rgw realm token in secondary site rgw spec
This was setting a field called "rgw_token" in the rgw spec
but this is not a real field in rgw specs. Instead we should
be setting "rgw_realm_token" which is what the field is
actually called.

Setting this nonexistent field causes the spec to be deleted
the first time cephadm needs to convert it from a json string
back into a python object (which happens whenever the module
restarts or the active mgr changes) which then causes all the
rgw daemons attached to the service to be removed

Fixes: https://tracker.ceph.com/issues/66824

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 95ae364325ddf3acc2ef9bac3e923fe9ca4ef6ba)